refactor fantasyx
